Use std::move where appropriate.
